Set up maps per buffer
If a pair is disabled for a filetype, we now skip mapping that pair for
that filetype, instead of checking whether it's disabled for the
filetype in the RHS of the mapping.

This also ensure that whether a pair is disabled for a filetype has no
effect on the commandline. Previously, the same check for whether a pair
is disabled was used in insert-mode and commandline-mode; now these are
separate checks, and filetypes do not affect commandline-mode.